move back methods to utils except define and find Strategy
