+++ packages/webdriverio/src/commands/element/$$.js
@@ -41,14 +41,12 @@ import { webdriverMonad } from 'webdriver'
 import { wrapCommand, runFnInFiberContext } from '@wdio/config'
 import merge from 'lodash.merge'
 
-import { getPrototype as getWDIOPrototype } from '../../utils'
+import { findElements, getBrowserObject, getPrototype as getWDIOPrototype } from '../../utils'
 import { elementErrorHandler } from '../../middlewares'
 import { ELEMENT_KEY } from '../../constants'
 import {
-    findElements,
-    getBrowserObject,
     getElementFromResponse
-} from '../../find-strategy'
+} from '../../utils'
 
 export default async function $$ (selector) {
     const res = await findElements.call(this, selector)
