+++ packages/webdriverio/src/commands/element/$.js
@@ -45,14 +45,12 @@ import { webdriverMonad } from 'webdriver'
 import { wrapCommand, runFnInFiberContext } from '@wdio/config'
 import merge from 'lodash.merge'
 
-import { getPrototype as getWDIOPrototype } from '../../utils'
+import { findElement, getBrowserObject, getPrototype as getWDIOPrototype } from '../../utils'
 import { elementErrorHandler } from '../../middlewares'
 import { ELEMENT_KEY } from '../../constants'
 import {
-    findElement,
-    getBrowserObject,
     getElementFromResponse
-} from '../../find-strategy'
+} from '../../utils'
 
 export default async function $ (selector) {
     const res = await findElement.call(this, selector)
