+++ packages/webdriverio/src/commands/element/selectByVisibleText.js
@@ -28,7 +28,7 @@
  *
  */
 
-import { getElementFromResponse } from '../../find-strategy'
+import { getElementFromResponse } from '../../utils'
 
 export default async function selectByVisibleText (text) {
     /**
